+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Target encoding for multiclass in CategoricalEncoding
 - Handle models that not support nan values
+- Internal use of RecursiveFeatureAddition
 
 ## [1.0.0] - 2025-01-03